Add a conversation cache
The id of the conversatino is cached for about a day while the bot is active. Otherwise if it has a record for that conversation, then a message is sent for that conversation. After each message to the bot, the conversation id gets refreshed to ensure that there will be no disconnects.